requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
arrow==1.2.3
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
defusedxml==0.7.1
Flask==2.3.2
frozenlist==1.4.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
openai==0.27.8
packaging==23.1
pyTenable==1.4.13
python-box==7.0.1
python-dateutil==2.8.2
requests==2.31.0
restfly==1.4.7
semver==3.0.1
six==1.16.0
tqdm==4.65.0
typing-extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.3.6
yarl==1.9.2
zipp==3.16.2
